push fixes and tiny makefile changes

+++ lib/libft/src/lib/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: fheaton- <fheaton-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 16:55:11 by fporto            #+#    #+#             */
-/*   Updated: 2025/08/27 20:45:44 by fheaton-         ###   ########.fr       */
+/*   Updated: 2025/08/29 23:24:18 by fheaton-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 char	*ft_strrchr(const char *s, int c)
 {
 	int		i;
-	size_t	len;
+	int		len;
 
 	i = 0;
 	len = ft_strlen(s);
 	if (c == '\0')
 		return ((char *)s + len);
-	i = (int)len - 1;
+	i = len - 1;
 	while (i >= 0)
 	{
 		if (s[i] == (unsigned char)c)
